requirements.txt
wagtail
django-appconf
psycopg2
djangorestframework
markdown
django-filter
